issues drawing to canvas
