+++ index.html
@@ -12,7 +12,7 @@
           width="800px"
           height="600px"></canvas>
 
-  <script src="./lib/enemies.js"></script>
+  <script src="./lib/index.js"></script>
 
 </body>
 </html>
