+++ lib/enemies.js
@@ -49,7 +49,10 @@ window.onload = function() {
   function drawEverything() {
     colorRect(0,0,canvas.width,canvas.height, 'black');
     colorCircle(ballX, ballY, 10, 'red')
-    colorCircle(ballX, 48, 5, 'red')
+    colorCircle(752, 52, 10, 'red')
+    // colorCircle(ballX, 48, 5, 'red')
+    // colorCircle(300, ballY, 5, 'red')
+    // colorCircle(ballY, ballX, 10, 'red')
   }
 
   function colorRect(x,y,width,height,drawColor){
